requirements.txt
streamlit==0.88.0
psycopg2==2.9.1
pandas==1.3.3
Pillow==8.4.0
numpy==1.21.2
easyocr==1.4.0
